+++ include/feature/signal_handler.h
@@ -1,8 +1,0 @@
-#ifndef FEATURE_SIGNAL_HANDLER_H
-#define FEATURE_SIGNAL_HANDLER_H
-
-#include <signal.h>
-
-void signal_handler(int signum);
-
-#endif