+++ src/commands/cd.cpp
@@ -1,4 +1,4 @@
-#include "cmds.h"
+#include "shell.h"
 
 #include <iostream>
 #include <filesystem>
@@ -20,13 +20,13 @@ static int cd_single(std::string_view path,
     } else if (std::filesystem::is_directory(current_path.append(path))) {
         return 0;
     } else {
-        std::cerr << '`' << path << "` is not a directory.\n";
+        fmt::print(stderr, "cd: {} is not a directory.\n", path);
         return 1;
     }
     return 0;
 }
 
-int cmds::cd(const std::vector<std::string>& arg, VariableManager& vars) {
+int Shell::cmd_cd(const std::vector<std::string>& arg) {
     if (arg.size() != 2) {
         return 1;
     }