+++ src/commands/set.cpp
@@ -1,4 +1,4 @@
-#include "cmds.h"
+#include "shell.h"
 
 #include <iostream>
 #include <filesystem>
@@ -6,9 +6,9 @@
 
 #include <fmt/format.h>
 
-int cmds::set(const std::vector<std::string>& arg, VariableManager& vars) {
+int Shell::cmd_set(const std::vector<std::string>& arg) {
     if (arg.size() != 3) {
-        std::cerr << "Argument size error.\n";
+        fmt::print(stderr, "Argument size error.\n");
         return 1;
     }